requirements.txt
Django==1.11
djangorestframework==3.6.3
psycopg2==2.7.1